+++ src/elements/mod.rs
@@ -1,3 +1,5 @@
 mod b_box;
+mod b_button;
 
 pub use b_box::*;
+pub use b_button::*;